+++ src/ts/AdUnits/VideoAdUnit.ts
@@ -3,7 +3,7 @@ import { AbstractAdUnit } from 'AdUnits/AbstractAdUnit';
 import { FinishState } from 'Constants/FinishState';
 import { Placement } from 'Models/Placement';
 import { Campaign } from 'Models/Campaign';
-import { AdUnitContainer } from 'AdUnits/Containers/AdUnitContainer';
+import { AdUnitContainer, ForceOrientation} from 'AdUnits/Containers/AdUnitContainer';
 import { Double } from 'Utilities/Double';
 import { Video } from 'Models/Video';
 import { Overlay } from 'Views/Overlay';
@@ -15,14 +15,14 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
 
     private static _progressInterval: number = 250;
 
+    protected _onShowObserver: any;
+    protected _onSystemKillObserver: any;
+    protected _onSystemInterruptObserver: any;
+
+    protected _options: any;
     private _video: Video;
     private _overlay: Overlay | undefined;
     private _deviceInfo: DeviceInfo;
-    private _options: any;
-
-    private _onShowObserver: any;
-    private _onSystemKillObserver: any;
-    private _onSystemInterruptObserver: any;
 
     constructor(nativeBridge: NativeBridge, container: AdUnitContainer, placement: Placement, campaign: Campaign, video: Video, overlay: Overlay, deviceInfo: DeviceInfo, options: any) {
         super(nativeBridge, container, placement, campaign);
@@ -42,7 +42,7 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
         this._onSystemKillObserver = this._container.onSystemKill.subscribe(() => this.onSystemKill());
         this._onSystemInterruptObserver = this._container.onSystemInterrupt.subscribe(() => this.onSystemInterrupt());
 
-        return this._container.open(this, true, !this._placement.useDeviceOrientationForVideo(), this._placement.disableBackButton(), this._options);
+        return this._container.open(this, true, true, this._placement.useDeviceOrientationForVideo() ? ForceOrientation.NONE : ForceOrientation.LANDSCAPE, this._placement.disableBackButton(), this._options);
     }
 
     public hide(): Promise<void> {
@@ -77,15 +77,7 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
         delete this._overlay;
     }
 
-    private hideChildren() {
-        const overlay = this.getOverlay();
-
-        if(overlay) {
-            overlay.container().parentElement!.removeChild(overlay.container());
-        }
-    };
-
-    private onShow() {
+    protected onShow() {
         if(this.isShowing() && this._video.isActive()) {
             if(this._nativeBridge.getPlatform() === Platform.IOS && IosUtils.hasVideoStallingApi(this._deviceInfo.getOsVersion())) {
                 if(this.getVideo().isCached()) {
@@ -99,18 +91,26 @@ export abstract class VideoAdUnit extends AbstractAdUnit {
         }
     }
 
-    private onSystemKill() {
+    protected onSystemKill() {
         if(this.isShowing()) {
             this.setFinishState(FinishState.SKIPPED);
             this.hide();
         }
     }
 
-    private onSystemInterrupt(): void {
+    protected onSystemInterrupt(): void {
         if(this.isShowing() && this._video.isActive()) {
             this._nativeBridge.Sdk.logInfo('Continuing Unity Ads video playback after interrupt');
             this._nativeBridge.VideoPlayer.play();
         }
     }
+
+    protected hideChildren() {
+        const overlay = this.getOverlay();
+
+        if(overlay) {
+            overlay.container().parentElement!.removeChild(overlay.container());
+        }
+    };
 
 }