+++ src/ts/EventHandlers/EndScreenEventHandlers.ts
@@ -9,6 +9,7 @@ import { DeviceInfo } from 'Models/DeviceInfo';
 import { IosUtils } from 'Utilities/IosUtils';
 import { PerformanceCampaign } from 'Models/PerformanceCampaign';
 import { StoreName } from 'Models/PerformanceCampaign';
+import { Diagnostics } from 'Utilities/Diagnostics';
 
 export class EndScreenEventHandlers {
 
@@ -18,21 +19,10 @@ export class EndScreenEventHandlers {
 
         nativeBridge.Listener.sendClickEvent(adUnit.getPlacement().getId());
 
-        if(campaign.getClickAttributionUrlFollowsRedirects()) {
-            sessionManager.sendClick(adUnit).then(response => {
-                const location = Request.getHeader(response.headers, 'location');
-                if(location) {
-                    nativeBridge.Intent.launch({
-                        'action': 'android.intent.action.VIEW',
-                        'uri': location
-                    });
-                } else {
-                    throw new Error('No location found');
-                }
-            });
+        sessionManager.sendClick(adUnit);
+        if(campaign.getClickAttributionUrl()) {
+            EndScreenEventHandlers.handleClickAttribution(nativeBridge, sessionManager, campaign);
         } else {
-            sessionManager.sendClick(adUnit);
-            nativeBridge.Sdk.logInfo(EndScreenEventHandlers.getAppStoreUrl(campaign, packageName));
             nativeBridge.Intent.launch({
                 'action': 'android.intent.action.VIEW',
                 'uri': EndScreenEventHandlers.getAppStoreUrl(campaign, packageName)
@@ -45,17 +35,10 @@ export class EndScreenEventHandlers {
 
         nativeBridge.Listener.sendClickEvent(adUnit.getPlacement().getId());
 
-        if(campaign.getClickAttributionUrlFollowsRedirects()) {
-            sessionManager.sendClick(adUnit).then(response => {
-                const location = Request.getHeader(response.headers, 'location');
-                if(location) {
-                    nativeBridge.UrlScheme.open(location);
-                } else {
-                    throw new Error('No location found');
-                }
-            });
+        sessionManager.sendClick(adUnit);
+        if(campaign.getClickAttributionUrl()) {
+            EndScreenEventHandlers.handleClickAttribution(nativeBridge, sessionManager, campaign);
         } else {
-            sessionManager.sendClick(adUnit);
             if(IosUtils.isAppSheetBroken(deviceInfo.getOsVersion()) || campaign.getBypassAppSheet()) {
                 nativeBridge.UrlScheme.open(EndScreenEventHandlers.getAppStoreUrl(campaign));
             } else {
@@ -77,6 +60,31 @@ export class EndScreenEventHandlers {
                 });
             }
         }
+    }
+
+    public static handleClickAttribution(nativeBridge: NativeBridge, sessionManager: SessionManager, campaign: PerformanceCampaign) {
+        const eventManager = sessionManager.getEventManager();
+        const platform = nativeBridge.getPlatform();
+
+        eventManager.clickAttributionEvent(sessionManager.getSession().getId(), campaign.getClickAttributionUrl(), campaign.getClickAttributionUrlFollowsRedirects()).then(response => {
+            const location = Request.getHeader(response.headers, 'location');
+            if(location) {
+                if(platform === Platform.ANDROID) {
+                    nativeBridge.Intent.launch({
+                        'action': 'android.intent.action.VIEW',
+                        'uri': location
+                    });
+                } else if(platform === Platform.IOS) {
+                    nativeBridge.UrlScheme.open(location);
+                }
+            } else {
+                Diagnostics.trigger('click_attribution_failed', {
+                    url: campaign.getClickAttributionUrl(),
+                    followsRedirects: campaign.getClickAttributionUrlFollowsRedirects(),
+                    response: response
+                });
+            }
+        });
     }
 
     public static onPrivacy(nativeBridge: NativeBridge, url: string): void {