+++ src/ts/Test/Integration/HappyTest.ts
@@ -53,7 +53,7 @@ describe('IntegrationTest', () => {
         DeviceInfo.setFreeSpace(10159440);
         DeviceInfo.setConnectionType('wifi');
         DeviceInfo.setNetworkType(0);
-        DeviceInfo.setNetworkOperator(24412);
+        DeviceInfo.setNetworkOperator('24412');
         DeviceInfo.setNetworkOperatorName('DNA');
         DeviceInfo.setHeadset(false);
         DeviceInfo.setDeviceVolume(1);
@@ -105,7 +105,7 @@ describe('IntegrationTest', () => {
         DeviceInfo.setFreeSpace(10159440);
         DeviceInfo.setConnectionType('wifi');
         DeviceInfo.setNetworkType(0);
-        DeviceInfo.setNetworkOperator(24412);
+        DeviceInfo.setNetworkOperator('24412');
         DeviceInfo.setNetworkOperatorName('DNA');
         DeviceInfo.setHeadset(false);
         DeviceInfo.setDeviceVolume(1);