+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# Wed Mar  1 13:25:48 UTC 2017
+
+* Remove split screen ad unit A/B test
+* Add hit areas for overlay buttons
+* Fix campaign request parameters
+
+# Mon Feb 27 13:12:02 UTC 2017
+
+* Fix click attribution
+* Simple validation of campaign asset URLs (videos, endscreen images)
+* Refactor test metadata fetching
+
+# Wed Feb 22 13:33:10 UTC 2017
+
+* Split screen ad unit A/B test
+
+# Tue Feb 21 11:31:53 UTC 2017
+
+* Send skip event to Kafka instead of comet
+* Send frameworkName and frameworkVersion parameters in ad requests and operative events
+
 # Mon Feb 20 10:52:10 UTC 2017
 
 * Restore invocations of iOS setAutomaticallyWaitsToMinimizeStalling API